pyproject.toml
[tool.poetry]
name = "rakam-systems"
version = "0.1.2"
description = ""
authors = ["MohammedJassem <mohammedjassemhlel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.1,<4.0"
beautifulsoup4 = "4.12.3"
channels = "4.0.0"
colorlog = "6.8.2"
drf-spectacular = "0.27.1"
faiss-cpu = "1.8.0.post1"
joblib = "1.4.2"
llama-parse = "0.4.9"
openai = "1.37.0"
panda = "0.3.1"
pandas = "1.3.5"
playwright = "1.45.0"
pre-commit = "3.5.0"
psycopg2-binary = "2.9.9"
pymupdf = "1.24.9"
python-dotenv = "1.0.1"
python-magic = "0.4.27"
requests = "2.31.0"
sentence-transformers = "2.6.1"
torch = "2.2.2"
tqdm = "4.66.4"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.2"
requests-mock = "^1.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
